+++ src/build_system.mli
@@ -232,6 +232,7 @@ module Rule : sig
 
   type t =
     { id      : Id.t
+    ; dir     : Path.t
     ; deps    : Deps.t
     ; targets : Path.Set.t
     ; context : Context.t option